M3: store Mode 6 scaling constant as a double
OBD2Mode6UnitAndScalingObject.constant was declared NSInteger, so the
UASID 0x39 entry (constant -327.68) was truncated to -327 at table
build time. scaleValue: already returns double and adds the constant
directly, so widening the storage type preserves the fractional
component without further changes to the call sites.

+++ Sources/LTSupportAutomotive/LTOBD2Mode6TestResult.m
@@ -123,17 +123,17 @@ typedef enum : NSUInteger {
 @property(strong,nonatomic,readonly) NSString* name;
 @property(strong,nonatomic,readonly) NSString* unit;
 @property(assign,nonatomic,readonly) double multiplier;
-@property(assign,nonatomic,readonly) NSInteger constant;
+@property(assign,nonatomic,readonly) double constant;
 @property(assign,nonatomic,readonly) BOOL s;
 @property(strong,nonatomic,readonly) NSString* formatString;
 
-+(instancetype)unitAndScalingObjectWithName:(NSString*)name unit:(NSString*)unit multiplier:(double)multiplier constant:(NSInteger)constant s:(BOOL)s;
++(instancetype)unitAndScalingObjectWithName:(NSString*)name unit:(NSString*)unit multiplier:(double)multiplier constant:(double)constant s:(BOOL)s;
 
 @end
 
 @implementation OBD2Mode6UnitAndScalingObject
 
-+(instancetype)unitAndScalingObjectWithName:(NSString*)name unit:(NSString*)unit multiplier:(double)multiplier constant:(NSInteger)constant s:(BOOL)s
++(instancetype)unitAndScalingObjectWithName:(NSString*)name unit:(NSString*)unit multiplier:(double)multiplier constant:(double)constant s:(BOOL)s
 {
     OBD2Mode6UnitAndScalingObject* obj = [[self alloc] init];
     obj->_name = name;
